+++ visualization/app/codeCharta/ui/ribbonBar/artificialIntelligence/suspiciousMetrics/suspiciousMetrics.component.spec.ts
@@ -17,8 +17,8 @@ describe("SuspiciousMetricsComponent", () => {
 	})
 
 	describe("badge", () => {
-		it("should show initially and hide on first click", async () => {
-			const { container } = await render(SuspiciousMetricComponent, {
+		it("should show initially and hide on first click, but show again when data has changed", async () => {
+			const { container, rerender } = await render(SuspiciousMetricComponent, {
 				excludeComponentDeclaration: true,
 				componentProperties: {
 					data: {
@@ -32,6 +32,44 @@ describe("SuspiciousMetricsComponent", () => {
 			expect(container.querySelector(".suspicious-metrics-badge")).not.toBe(null)
 
 			userEvent.click(screen.getByTitle("Open Suspicious Metrics Panel"))
+			expect(container.querySelector(".suspicious-metrics-badge")).toBe(null)
+
+			rerender({
+				data: {
+					analyzedProgrammingLanguage: "ts",
+					unsuspiciousMetrics: ["rloc"],
+					suspiciousMetricSuggestionLinks: [],
+					untrackedMetrics: []
+				}
+			})
+			expect(container.querySelector(".suspicious-metrics-badge")).not.toBe(null)
+		})
+
+		it("should show initially and hide on first click, but not show again when new data has same values", async () => {
+			const { container, rerender } = await render(SuspiciousMetricComponent, {
+				excludeComponentDeclaration: true,
+				componentProperties: {
+					data: {
+						analyzedProgrammingLanguage: "ts",
+						unsuspiciousMetrics: ["rloc", "mcc"],
+						suspiciousMetricSuggestionLinks: [],
+						untrackedMetrics: []
+					}
+				}
+			})
+			expect(container.querySelector(".suspicious-metrics-badge")).not.toBe(null)
+
+			userEvent.click(screen.getByTitle("Open Suspicious Metrics Panel"))
+			expect(container.querySelector(".suspicious-metrics-badge")).toBe(null)
+
+			rerender({
+				data: {
+					analyzedProgrammingLanguage: "ts",
+					unsuspiciousMetrics: ["rloc", "mcc"],
+					suspiciousMetricSuggestionLinks: [],
+					untrackedMetrics: []
+				}
+			})
 			expect(container.querySelector(".suspicious-metrics-badge")).toBe(null)
 		})