+++ visualization/app/codeCharta/ui/ribbonBar/artificialIntelligence/suspiciousMetrics/suspiciousMetrics.component.ts
@@ -1,4 +1,5 @@
-import { Component, Inject, Input } from "@angular/core"
+import { Component, Inject, Input, OnChanges, SimpleChanges } from "@angular/core"
+import isEqual from "lodash.isequal"
 import { Store } from "../../../../state/angular-redux/store"
 import { defaultMapColors, setMapColors } from "../../../../state/store/appSettings/mapColors/mapColors.actions"
 import { setAreaMetric } from "../../../../state/store/dynamicSettings/areaMetric/areaMetric.actions"
@@ -13,13 +14,20 @@ import { MetricSuggestionParameters } from "../selectors/util/suspiciousMetricsH
 	selector: "cc-suspicious-metrics",
 	template: require("./suspiciousMetrics.component.html")
 })
-export class SuspiciousMetricComponent {
+export class SuspiciousMetricComponent implements OnChanges {
 	@Input() data: Pick<
 		ArtificialIntelligenceData,
 		"analyzedProgrammingLanguage" | "unsuspiciousMetrics" | "suspiciousMetricSuggestionLinks" | "untrackedMetrics"
 	>
+	hideBadge = false
 
 	constructor(@Inject(Store) private store: Store) {}
+
+	ngOnChanges(changes: SimpleChanges): void {
+		if (changes.data && !isEqual(changes.data.previousValue, changes.data.currentValue)) {
+			this.hideBadge = false
+		}
+	}
 
 	applySuspiciousMetric(metric: MetricSuggestionParameters, markOutlier: boolean) {
 		this.store.dispatch(setAreaMetric(AREA_METRIC))